force json data to UTF-8 in tree history test
JSON.parse() will crash with an Encoding::InvalidByteSequenceError if it
tries to parse a UTF-8 character in a string that Ruby does not consider
to be UTF-8.

Our test data contains UTF-8 characters, but prior to this commit, the
encoding of our test data was implicit (Encoding.default_external). If
the LANG environment variable was not UTF-8, then Ruby would not
consider the test data to have a UTF-8 encoding, and JSON.parse() would
crash.

The simplest way to fix this to explicitly set the encoding of the data
that we pass to JSON.parse(). This removes the dependency on
Encoding.default_external and the LANG environment variable.