tests: add "refs" dir to dolt-test-repo.git
The Rugged gem would not read the dolt-test-repo.git location unless a
"refs" directory is present in the bare fixture repository.

Add the "refs" directory and a .gitkeep file so we always have this
directory in all Git clones.

This resovles a test suite failure on CentOS 6 with Ruby 1.9.3 SCL and
the Rugged v0.19.0 gem.